v1.0 of exploratory_analisis.py with the implementation of utils.DE as a module for making the diferential expression analysis, same implementation was made with utils_DW.py for Download.py
